Moving settings into Convars so can be altered via console.
The settings were previously just hardcoded in and set into global
variables. If we set them as ConVars we can then set them via the
console.

For some reason the gameDescription.txt settings are not reflected on to
these Convars so it is probably best if you do not show them to people
as they will believe they do something!